pyproject.toml
[project]
name = "autonomous-code-review-agent-with-dynamic-replanning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "gitpython>=3.1.46",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-google-genai>=4.2.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
